feat(tartarus): move script structure into talaria
