+++ tartarus/Cargo.toml
@@ -14,8 +14,8 @@ log = "0.4"
 env_logger = "0.11.8"
 rustyline = { version = "15.0.0", features = ["derive"]}
 anyhow = "1.0"
-toml = "0.8"
 sysinfo = "0.34.2"
 strum = "0.27.1"
 structz-macros = "0.2.0"
 figment = {version = "0.10", features = ["toml", "env"]}
+toml = "0.8.22"
